requirements.txt
asgiref==3.2.10
Django==3.1
django-autoslug==1.9.8
django-crispy-forms==1.11.2
django-phone-field==1.8.1
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
